Implement backend loading logic
Try to use xdg-desktop-portal when possible, but also watch for its
DBus service to appear/disappear so we can switch between backends
on runtime. This is usefull when we start an app without xdp not
being ready and later once it becomes ready we can use it and reload
app configuration.